+++ tasks/src/app/classes/modal.ts
@@ -1,0 +1,18 @@
+import { ICourse } from '../interfaces/course';
+
+export class Modal {
+    public isOpen: boolean = false;
+    public closeFn?: () => void;
+    public course?: ICourse = { 
+        id: null,
+        name: '',
+        date: new Date(),
+        length: null,
+        description: '', 
+        authors: {
+            id: null,
+            name: '',
+        },
+        isTopRated: false
+    };
+}
