+++ tasks/src/app/components/search/search.component.html
@@ -1,4 +1,4 @@
 <div class="search-panel">
-  <input type="text" placeholder="Text to search"  class="search-input" [(ngModel)]="inputValue">
+  <input type="text" placeholder="Text to search" class="search-input" [(ngModel)]="inputValue">
   <button class="search" (click)='getInputText(inputValue)'>Search</button>
 </div>
